refactor(exception): centralize HttpMessageNotReadableException handling
- move exception handling from AuthControllerAdvice to ControllerExceptionHandler
- add logger initialization in ControllerExceptionHandler

refactor(dto): relocate CreateEmployeeRequest to employee request package

- migrate CreateEmployeeRequest to org.hyperskill.phonebook.dtos.request.employee
- update references across the codebase

fix(employee): handle nullable fields in CreateEmployeeRequest

- set default values for nullable fields in CreateEmployeeRequest
- use orEmpty() method to prevent null issues during Employee creation
