+++ src/main/kotlin/org/hyperskill/phonebook/controller/EmployeeController.kt
@@ -2,23 +2,15 @@ package org.hyperskill.phonebook.controller
 
 
 import jakarta.validation.Valid
-import org.hyperskill.phonebook.dtos.CreateEmployeeRequest
 import org.hyperskill.phonebook.dtos.UpdateEmployeeRequest
+import org.hyperskill.phonebook.dtos.request.employee.CreateEmployeeRequest
 import org.hyperskill.phonebook.model.Employee
 import org.hyperskill.phonebook.service.EmployeeServices
 import org.springframework.data.domain.Page
 import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
-import org.springframework.web.bind.annotation.GetMapping
-import org.springframework.web.bind.annotation.PostMapping
-import org.springframework.web.bind.annotation.RequestBody
-import org.springframework.web.bind.annotation.RequestMapping
-import org.springframework.web.bind.annotation.RequestParam
-import org.springframework.web.bind.annotation.RestController
-import org.springframework.web.bind.annotation.PutMapping
-import org.springframework.web.bind.annotation.DeleteMapping
-import org.springframework.web.bind.annotation.PathVariable
-import java.util.UUID
+import org.springframework.web.bind.annotation.*
+import java.util.*
 
 
 @RestController
