+++ src/main/kotlin/org/hyperskill/phonebook/dtos/request/employee/CreateEmployeeRequest.kt
@@ -1,4 +1,4 @@
-package org.hyperskill.phonebook.dtos
+package org.hyperskill.phonebook.dtos.request.employee
 
 import jakarta.validation.constraints.Email
 import jakarta.validation.constraints.NotBlank
@@ -6,16 +6,16 @@ import java.util.*
 
 data class CreateEmployeeRequest(
     @field:NotBlank(message = "Name must not be blank")
-    val name: String,
+    val name: String? = null,
 
     @field:NotBlank(message = "Position must not be blank")
-    val position: String,
+    val position: String? = null,
 
     @field:NotBlank(message = "Phone must not be blank")
-    val phone: String,
+    val phone: String? = null,
 
     @field:Email(message = "Email must be a valid email address")
-    val email: String?,
+    val email: String? = null,
 
-    val departmentId: UUID?
+    val departmentId: UUID? = null
 )
