+++ src/main/kotlin/org/hyperskill/phonebook/exception/AuthControllerAdvice.kt
@@ -5,7 +5,6 @@ import org.hyperskill.phonebook.dtos.response.ErrorResponse
 import org.slf4j.Logger
 import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
-import org.springframework.http.converter.HttpMessageNotReadableException
 import org.springframework.security.authentication.BadCredentialsException
 import org.springframework.web.bind.MethodArgumentNotValidException
 import org.springframework.web.bind.annotation.ControllerAdvice
@@ -39,21 +38,6 @@ class AuthControllerAdvice {
                 statusCode = 400,
                 errorMessage = HttpStatus.BAD_REQUEST.reasonPhrase,
                 description = ex.bindingResult.fieldErrors.firstOrNull()?.defaultMessage ?: ex.message
-            ),
-            HttpStatus.BAD_REQUEST
-        )
-    }
-
-    @ExceptionHandler(HttpMessageNotReadableException::class)
-    fun handleException(
-        ex: HttpMessageNotReadableException
-    ): ResponseEntity<ErrorResponse> {
-        LOGGER.error("Error parsing request body", ex)
-        return ResponseEntity(
-            ErrorResponse(
-                statusCode = 400,
-                errorMessage = "Invalid Request Format",
-                description = ex.message
             ),
             HttpStatus.BAD_REQUEST
         )
