+++ src/main/kotlin/org/hyperskill/phonebook/service/EmployeeService.kt
@@ -1,6 +1,8 @@
 package org.hyperskill.phonebook.service
 
-import org.hyperskill.phonebook.dtos.CreateEmployeeRequest
+import jakarta.persistence.EntityNotFoundException
+import org.hyperskill.phonebook.dtos.UpdateEmployeeRequest
+import org.hyperskill.phonebook.dtos.request.employee.CreateEmployeeRequest
 import org.hyperskill.phonebook.model.Employee
 import org.hyperskill.phonebook.repository.DepartmentRepository
 import org.hyperskill.phonebook.repository.EmployeeRepository
@@ -8,9 +10,7 @@ import org.springframework.data.domain.Page
 import org.springframework.data.domain.PageRequest
 import org.springframework.data.repository.findByIdOrNull
 import org.springframework.stereotype.Service
-import org.hyperskill.phonebook.dtos.UpdateEmployeeRequest
-import jakarta.persistence.EntityNotFoundException
-import java.util.UUID
+import java.util.*
 
 @Service
 class EmployeeServices(
@@ -26,9 +26,9 @@ class EmployeeServices(
     fun store(createEmployeeRequest: CreateEmployeeRequest): Employee {
         val department = createEmployeeRequest.departmentId?.let { departmentRepository.findByIdOrNull(it) }
         val employee = Employee(
-            name = createEmployeeRequest.name,
-            position = createEmployeeRequest.position,
-            phone = createEmployeeRequest.phone,
+            name = createEmployeeRequest.name.orEmpty(),
+            position = createEmployeeRequest.position.orEmpty(),
+            phone = createEmployeeRequest.phone.orEmpty(),
             email = createEmployeeRequest.email,
             department = department
         )
